ArrowStreamWriter doesn't work on a non-seekable stream
Allow ArrowStreamWriter to write to a non-seekable stream, like a network stream.

+++ csharp/src/Apache.Arrow/ArrowBuffer.Builder.cs
@@ -111,7 +111,7 @@ namespace Apache.Arrow
 
             public ArrowBuffer Build(MemoryPool pool = default)
             {
-                var length = BitUtility.RoundUpToMultipleOf64(_buffer.Length);
+                var length = (int)BitUtility.RoundUpToMultipleOf64(_buffer.Length);
                 var memoryPool = pool ?? MemoryPool.Default.Value;
                 var memory = memoryPool.Allocate(length);
